PAMI Code Updated
Recently there was an update from the PAMI library on Oct 11, so the block of code related to generating a transactional database had to be updated in order to function correctly. If there are some students that have installed PAMI before that date please update the PAMI library in python, or use the old code that is commented in said block of code. This is right after exercise 16.